Refactor, moving inner classes from AutomatonTest to top-level in main/ org.apache.calcite.runtime
